+++ packages/components-providers/src/ScrollLock/ScrollLockProvider.tsx
@@ -24,78 +24,17 @@
 
  */
 
-import noop from 'lodash/noop'
-import React, { FC, useRef, useMemo } from 'react'
-import { ScrollLockContext } from './ScrollLockContext'
-import { activateScrollLock, getActiveScrollLock, ScrollLockMap } from './utils'
+import React, { createContext, FC } from 'react'
+import { TrapStackContextProps, TrapStackProvider } from '../TrapStack'
+import { activateScrollLock } from './utils'
 
-export const ScrollLockProvider: FC = ({ children }) => {
-  // stores all available scroll locks
-  // (map of ids to elements where scrolling is to be allowed)
-  const registeredLocksRef = useRef<ScrollLockMap>({})
-  // stores the current element where scrolling is allowed
-  // null if no scroll lock is active
-  const activeLockRef = useRef<HTMLElement | null>(null)
-  // stores the callback to remove the scroll event listener & restore body styles
-  const deactivateRef = useRef<() => void>(noop)
+export const ScrollLockContext = createContext<TrapStackContextProps>({})
+ScrollLockContext.displayName = 'ScrollLockContext'
 
-  // useMemo: this component probably won't re-render much, but if it does
-  // let's not cause unnecessary diffs from useScrollLock in Dialogs and Popovers
-  // since those components can be finicky if re-rendered at the wrong moment
-  const value = useMemo(() => {
-    function getLock(id?: string): HTMLElement | null {
-      const registeredLocks = registeredLocksRef.current
-      return id
-        ? registeredLocks[id] || null
-        : getActiveScrollLock(registeredLocks)
-    }
-
-    function enableCurrentLock() {
-      const newElement = getLock()
-      if (newElement !== activeLockRef.current) {
-        // Disable the existing scroll lock and update the element
-        activeLockRef.current = newElement
-        deactivateRef.current()
-        // If there's a new element, activate a new scroll lock
-        if (newElement) {
-          deactivateRef.current = activateScrollLock(newElement)
-        }
-      }
-    }
-
-    function disableCurrentLock() {
-      deactivateRef.current()
-      deactivateRef.current = noop
-      activeLockRef.current = null
-    }
-
-    function addLock(id: string, element: HTMLElement) {
-      registeredLocksRef.current[id] = element
-      enableCurrentLock()
-    }
-
-    function removeLock(id: string) {
-      const existingLock = getLock(id)
-      if (existingLock) {
-        const registeredLocks = registeredLocksRef.current
-        delete registeredLocks[id]
-        enableCurrentLock()
-      }
-    }
-
-    return {
-      activeLockRef,
-      addLock,
-      disableCurrentLock,
-      enableCurrentLock,
-      getLock,
-      removeLock,
-    }
-  }, [])
-
-  return (
-    <ScrollLockContext.Provider value={value}>
-      {children}
-    </ScrollLockContext.Provider>
-  )
-}
+export const ScrollLockProvider: FC = (props) => (
+  <TrapStackProvider
+    activate={activateScrollLock}
+    context={ScrollLockContext}
+    {...props}
+  />
+)