+++ packages/components/src/Dialog/useDialog.tsx
@@ -154,14 +154,8 @@ export const useDialog = ({
       ? controlledSetOpen
       : setUncontrolledIsOpen
 
-  const {
-    callbackRef: focusRef,
-    disable: disableFocusTrap,
-    enable: enableFocusTrap,
-    isEnabled: focusTrapEnabled,
-    trapRef: focusTrapRef,
-  } = useFocusTrap(isOpen)
-  const [, portalRef] = useScrollLock(focusRef)
+  const [, focusRef] = useFocusTrap()
+  const [, portalRef] = useScrollLock({ ref: focusRef })
 
   const handleOpen = () => setOpen(true)
 
@@ -177,10 +171,6 @@ export const useDialog = ({
     <DialogContext.Provider
       value={{
         closeModal: handleClose,
-        disableFocusTrap,
-        enableFocusTrap,
-        focusTrapEnabled,
-        focusTrapRef,
       }}
     >
       <Portal ref={portalRef}>